Refactor - dbmanager.py amended to remove need for second flask application instance
A wrapper function was previously used to bind the close_connection method to the app_teardown
attribute of the Flask application object, this would raise an exception since at the time on import
of the dbmanager module the Flask application object had not been passed through to the module this
app_teardown attribute could not be accessed.

This was temporarily fixed previously via the use of a second application instance which
was created within the dbmanager.py module to a variable which was reassigned later to
the flask application object passed from application.py.

Instead of setting the close_connection callback function via the use of a wrapper function, it is
now done by passing the callback method directly within the DBManager class init method. The creation
of a second Flask application object was then removed as it was redundant.